Make the constraints field based
This means that there can be circular dependencies between items, as
long as the fields are not circularly dependent. Lets "earlier" items
refer to results of later items as long as they form a DAG.

For example if there were two labels:
            T1
     .------------.
   L1| Short Label|R2
     '------------'
            B1
            T2
     .----------------.
   L2| Very Long Label|R2
     '----------------'
            B2

If it was desired to have the right edge of the short label line
up with the right edge of the long label, but have the long label
laid out below the short label this could be achieved with the
following constraints:

L1 = WLeft 10.
R1 = IRight (label2, 0.)
T1 = WTop 10.
B1 = PreferredH

L2 = WLeft 10.
R2 = PreferredW
T2 = IBottom (label1, 10.)
B2 = PreferredH

So label2 references label1 and vice versa, but they do so in
different fields, and those fields are not circular. This means
all values can be computed without issue.