Creating a more "legit" application, filesystem browser
Need a lot more infrastructure work to make composing elements nicer.
Want to be able to trigger things based on events inside other widgets.
For example: hitting "Enter" on the textbox. Would also be nice to be
able to assign shortcuts/hot-keys to various widgets to jump to them
faster. Also need to have a nicer way to display what widget is
focused.