Draw only what changed, but with bugs!
Problem is rendering doesn't ignore widgets that are in the middle
of being initialized. This means that a label that isn't attached to
a layout can have setText called which invalidates the widget which
then draws it whever it currently is. Most like in the top left. This
is incorrect, it's draw request should be ignored until it should
actually be "on-screen".